Added some "better" authentication middleware
It still sucks, but it is probably the best we are going to get with our own implementation, since we can't figure out how to use the proper middleware.

+++ Business/IUserRepository.cs
@@ -13,8 +13,8 @@ namespace Business
 
         void DeleteUserToken(string email);
 
-        long? GetUserIdFromToken(string token);
+        int? GetUserIdFromToken(string token);
 
-        List<GroupModel> GetGroupsForUser(long userId);
+        List<GroupModel> GetGroupsForUser(int userId);
     }
 }
